Class_Activity_2/1093589207/requirements.txt
flask
flask_sqlalchemy
flask_cors
requests
selenium
fpdf2